requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17.2
APScheduler==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
clarifai==2.6.2
click==8.0.1
colorama==0.4.4
configparser==3.8.1
cryptography==35.0.0
discord.py==1.7.3
distlib==0.3.2
distro==1.6.0
docopt==0.6.2
EasyProcess==0.3
et-xmlfile==1.1.0
filelock==3.0.12
future==0.18.2
greenlet==1.1.1
grpcio==1.41.0
gspread==3.7.0
html2text==2020.1.16
httplib2==0.19.1
isort==5.6.4
jsonschema==2.6.0
lazy-object-proxy==1.4.3
lxml==4.6.3
matrix-client==0.4.0
mccabe==0.6.1
MeaningCloud-python==2.0.0
multidict==5.2.0
numpy==1.21.2
oauth2client==4.1.3
oauthlib==3.1.1
openpyxl==3.0.7
pefile==2021.9.3
pipreqs==0.4.11
platformdirs==2.0.2
plyer==2.0.0
praw==7.5.0
prawcore==2.3.0
protobuf==3.17.3
psycopg2-binary==2.9.1
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.1
pywin32-ctypes==0.2.0
rsa==4.7.2
six==1.16.0
toml==0.10.1
tqdm==4.62.3
tzlocal==2.1
urllib3==1.26.5
webdriverdownloader==1.1.0.3
websocket-client==1.2.1
Werkzeug==2.0.1
wrapt==1.12.1

